Rename and merge test files
